+++ CHANGELOG.md
@@ -119,6 +119,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.14] - 2026-06-08
+
+### Docs
+- Update README.md
+
+### Test
+- Update tests/test_cqrs_drivers.py
+- Update tests/test_hillm_uri.py
+
+### Other
+- Update .env.ollama
+- Update project/planfile-tickets.yaml
+
 ## [0.4.13] - 2026-06-08
 
 ### Docs